feat(milestone): 🚀 First successful mainnet build and deployment
- Backend canister deployed: kzec7-4iaaa-aaaaj-qngpa-cai
- Frontend canister deployed: k6fel-rqaaa-aaaaj-qngpq-cai
- Complete build pipeline working
- Authentication system integrated
- Genesis sequence implemented
- Matrix UI theme completed

Note: Minor hang at anima/1 to be addressed in next iteration

+++ src/ai/types.rs
@@ -4,14 +4,14 @@ use std::collections::HashMap;
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct EmotionalAnalysis {
-    pub valence: f32,
-    pub arousal: f32,
-    pub dominance: f32,
+    pub valence: f64,
+    pub arousal: f64,
+    pub dominance: f64,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct MemoryImpact {
-    pub intensity: f32,
-    pub relevance: f32,
-    pub trait_impacts: HashMap<String, f32>,
+    pub intensity: f64,
+    pub relevance: f64,
+    pub trait_impacts: HashMap<String, f64>,
 }
